Only delete the target file when it is in fact a file - don't delete directories.
git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@269282 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/ant/Project.java
@@ -755,7 +755,7 @@ public class Project {
         if (overwrite || !destFile.exists() ||
             destFile.lastModified() < sourceFile.lastModified()) {
 
-            if (destFile.exists()) {
+            if (destFile.exists() && destFile.isFile()) {
                 destFile.delete();
             }
